refactor: calculateResult 함수를 LottoMachine 클래스로 옮기고 멤버 메서드로 변경
